// client/src/UI/common/ui-button.d.ts
import React from 'react';
import { BoxProps, SxProps } from '@mui/material';
import { ImageEntry } from '../hooks/use-image';
type AtlasEntry = {
    texture: string;
    left: number;
    top: number;
    width: number;
    height: number;
};
type Props = {
    onClick?: React.MouseEventHandler<HTMLDivElement>;
    children?: React.ReactNode;
    className?: string;
    icon?: React.ReactNode;
    normal?: AtlasEntry;
    pressed?: AtlasEntry;
    hover?: AtlasEntry;
    disabled?: AtlasEntry;
    isDisabled?: boolean;
    buttonName?: string;
    sx?: SxProps;
    textSx?: SxProps;
    text?: string;
    scale?: number;
    textFontSize?: string;
    selected?: boolean;
    stone?: boolean;
    entrySx?: (entry: ImageEntry) => SxProps;
} & BoxProps;
export declare const UiButtonComponent: React.FC<Props>;
export {};